Split messages to reason and detail parts
* Move DOMImplmementation binding under src/binding
* Fix bug of PutForwards
* Implement restrict double binding

+++ src/binding/BlobBinding.cpp
@@ -28,7 +28,7 @@ static ESValue sizeGetterFunction(ESVMInstance* instance)
 {
     GENERATE_THIS_AND_CHECK_TYPE(Blob);
     // Declare native value (empty when type is void)
-    double result;
+    uint64_t result;
     result = originalObj->size();
     // Return ESValue from native value
     return ESValue(result);
